backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
httpx
beautifulsoup4
pydantic-settings
python-multipart
apscheduler
python-dotenv
pytest
pytest-asyncio
pytest-cov
